// Qanat.Web/src/app/shared/generated/model/geography-with-bounding-box-dto.ts
/**
 * Qanat.API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BoundingBoxDto } from '././bounding-box-dto';

import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class GeographyWithBoundingBoxDto { 
    GeographyID?: number;
    GeographyName?: string;
    GeographyDisplayName?: string;
    BoundingBox?: BoundingBoxDto;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface GeographyWithBoundingBoxDtoForm { 
    GeographyID?: FormControl<number>;
    GeographyName?: FormControl<string>;
    GeographyDisplayName?: FormControl<string>;
    BoundingBox?: FormControl<BoundingBoxDto>;
}

export class GeographyWithBoundingBoxDtoFormControls { 
    public static GeographyID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static GeographyName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static GeographyDisplayName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static BoundingBox = (value: FormControlState<BoundingBoxDto> | BoundingBoxDto = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<BoundingBoxDto>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}